+++ src/components/HeaderBar.js
@@ -1,19 +1,15 @@
-import { Container, Navbar, Section } from 'react-bulma-components';
+import { Container, Navbar } from 'react-bulma-components';
 import { Link, useNavigate } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import { useRecoilState, useSetRecoilState } from 'recoil';
 
-import AnimatedIcon from './icons/AnimatedIcon';
-import { BorderRadius } from '../styles';
-import Buttons from './Buttons';
-import { Colors } from '../styles/colors';
 import { FontSize } from '../styles/font';
+import HamburgerSideBar from './HamburgerSideBar';
 import HeaderSwitcher from './HeaderSwitcher';
 import { Media } from '../styles/media';
 import { Shadows } from '../styles/shadow';
 import Swal from 'sweetalert2';
 import { headerClickState } from '../state/headerState';
-import { keyframes } from '@emotion/react';
 import { loginState } from '../state/loginState';
 import styled from '@emotion/styled';
 
@@ -28,14 +24,6 @@ const HeaderBar = () => {
 
   const handleBurgerClick = () => {
     setVisible(!visible);
-  };
-
-  const handleBurgerCancleClick = () => {
-    setVisible(false);
-  };
-
-  const handlePageButtonClick = path => {
-    navigate(path);
   };
 
   const handleLinkClick = () => {
@@ -118,61 +106,16 @@ const HeaderBar = () => {
           )}
         </HeaderMenu>
       </Container>
-
       {visibleState && (
-        <HamburgerContent disappear={visible}>
-          <Container className="has-text-center p-3">
-            <HamburgerHeader>
-              <HamburgerBackButton size={'large'} onClick={handleBurgerCancleClick} />
-            </HamburgerHeader>
-            <Container>
-              <Section>
-                <p className="title">로고</p>
-                <DevideLine />
-                <DevideLine space="micro" color="none" />
-                <p className="subtitle">당신의 LinkyWay를 걸어봐요</p>
-                <p style={{ opacity: 0.7 }}>여기에 뭘 써야 할지 모르겠지만 시작해바요</p>
-                <DevideLine space="medium" color="none" />
-                {!login ? (
-                  <>
-                    <Buttons onClick={() => handlePageButtonClick('/login')}>로그인</Buttons>
-                    <DevideLine space="micro" color="none" />
-                    <Buttons onClick={() => handlePageButtonClick('/join')}>회원가입</Buttons>
-                  </>
-                ) : (
-                  <>
-                    <Buttons colortype={'sub'} onClick={handleLogoutClick}>
-                      로그아웃
-                    </Buttons>
-                    <DevideLine space="micro" color="none" />
-                  </>
-                )}
-              </Section>
-            </Container>
-          </Container>
-        </HamburgerContent>
+        <HamburgerSideBar
+          onCancle={handleBurgerClick}
+          onLogout={handleLogoutClick}
+          visible={visible}
+        />
       )}
     </StyledNavBar>
   );
 };
-
-const slideLeft = keyframes`
-from {
-  transform: translateX(500px);
-}
-to {
-  transform: translateX(0px);
-}
-`;
-
-const slideRight = keyframes`
-  from {
-    transform: translateX(0px);
-  }
-  to {
-    transform: translateX(500px);
-  }
-`;
 
 const StyledNavBar = styled(Navbar)`
   margin-top: -10px;
@@ -203,52 +146,6 @@ const StyledA = styled.a`
   text-decoration: 1px solid underline;
   text-underline-offset: ${FontSize.normal};
   opacity: 0.7;
-`;
-
-const HamburgerContent = styled.div`
-  z-index: 3;
-  position: absolute;
-  background-color: ${Colors.backgroundForm};
-  box-shadow: ${Shadows.card};
-  text-align: center;
-  justify-content: center;
-  align-content: center;
-  justify-content: center;
-  top: 0;
-  right: 0;
-  @media ${Media.desktop} {
-    display: none;
-  }
-  @media ${Media.tablet} {
-    width: 60%;
-  }
-  @media ${Media.mobile} {
-    width: 100%;
-  }
-  background-color: ${Colors.backgroundForm};
-  animation-duration: 0.25s;
-  animation-timing-function: ease-out;
-  animation-name: ${props => (props.disappear ? slideLeft : slideRight)};
-`;
-
-const HamburgerHeader = styled.div`
-  text-align: left;
-`;
-
-const HamburgerBackButton = styled(AnimatedIcon.RightArrow)`
-  :hover {
-    border: 2px solid ${Colors.mainFirst};
-    border-radius: ${BorderRadius.button};
-    cursor: pointer;
-  }
-`;
-
-const DevideLine = styled.hr`
-  background-color: ${props => (!props.color ? Colors.mainFirst : props.color)};
-  opacity: 0.6;
-  box-shadow: ${props => (!props.color ? Shadows.button : 'none')};
-  margin-top: ${props => (!props.space ? FontSize.huge : FontSize[props.space])};
-  margin-bottom: ${props => (!props.space ? FontSize.huge : FontSize[props.space])};
 `;
 
 const CenterSwitchContainer = styled(StyledLogo)`
